backend/requirements.txt
Django==5.0.0
djangorestframework==3.14.0
django-cors-headers==4.3.0
#django-simple-jwt
psycopg2-binary==2.9.9
celery==5.3.6
redis==5.0.1
pillow==10.1.0
python-dotenv==1.0.0
django-filter==23.5
celery
redis
WeasyPrint